backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
yfinance==0.2.55
pandas==2.2.3
numpy==2.2.1
